providers/jules/requirements.txt
pandas
requests
pyyaml
python-slugify
jinja2
python-dotenv
openpyxl
